Fixed issue with AD deal request withdrawl logic
Looks like a bool was named incorrectly, leading to some inverted logic.
The bug would cause valid deals being withdrawl by the AI and invalid
deals no withdrawn.

The bug had been addressed by the time I got around to submitting the fix but I am not sure if the fix that was checked in is the best solution since I thought it would result in bad deals being presented to the player.

I have an upcoming pull request that does actually remove this condition but also implements logic to try to adjust the deal so that it is acceptable before presenting.