Enabled Events in MP
Added code to send Event related messages to remote players. However, this is done via the foul means of hijacking the DiplomacyFromUI message and giving it a dual purpose. Event messages are have a flag encoded into one of the arguments and this is detected and handled appropriatly on the clients.

Seems to work fine but maybe some events are problematic as I wouldn't have encountered all variations of mechanisms in limited testing.

Added code is nothing fancy and can be improved if we want to send more
of messages in this manner.