refreshing some cached data on make peace
This data depends on the warring state but is only updating in the event of war being declared. It seems to me it should be updated on peace too, keeping in mind that the parties could still be at war with other players.

Also renamed arg since it is being called for peace now too.

I have no evidence that this helps matters.