Added second autosave point and save events
Due to the barrier on the finshing of ai processing, it looks like it might be safe to slot in a second autosave point after AI has finished but before humans.

Unfortunately, saving at this point can result in a slightly unpleasant stuttering of the game. Nothing too bad but some may not like it and not want it enabled. Also because of the was MP saves work the post ai save will be used for manually saved games also.

In order to give people better choice as to when to have the post ai save enabled I have also added a couple of GameEvents which allow the saves to be enabled/disabled on an indivual basis, turn by turn.

It is quite barebones and actually removed features since I am wary of enlarging the API. I have implemented a separate Lua UI mod that gives fine grained control over saves as a proof of concept, at least. It is pretty nasty and wouldn't be included in VP in the current state. It is something that is probably best left to the mod modders and now the tools are there.