Forcing and Invalidation of TacticalMaps
TacticalMaps were going out of sync. To try to avoid this I am forcing a refresh in CvPlayer::doTurnPostDiplomacy. Unfortunately, I am not sure why the maps were going out of sync.

This change is very heavy handed but does appear to help prevent
desyncs. It can be cleaned up and optimized but right now I just want to
prevent desyncs and don't want to worry about something not being
updated when it needed to be.

At any rate, it does look like they were being updated too early by CvPlayer::UpdateCityThreadCriteria. A comment in the TacticalMap class says that refreshing should be done after danger plots are updated in CvPlayer::doTurnPostDiplomacy - CvPlayer::UpdateCityThreatCriteria is called just before that.

I assume that it is desirable to have the TacticalMap refreshed during ai_unitUpdate, so I also implemented a means of invalidating the cache so that the map is updated then if need be...unless of course something else updates it beforehand.

This is just hanging together and could benefit from further thought but I am not familiar enough at the moment. Is it worth considering forcing an update in ai_unitMoves instead? Maybe just the invalidating is enough. Maybe just re-ordering UpdateCityThreatCriteria would be enough?

Now forcing TacticalMap refresh for all alive players after war/peace. This is almost certainly over the top and could be more targetted but for the moment I just want to fix the desyncs. Ultimately the caching should be improved so that it is invalidated appropriately and also refreshed whenever required.

Also updated but commented out code from CvTeam that was refreshing just the parties involved. This may be all that needs to be done but I (at least) am willing to take the extra processing time over a desync.