Fix make Trade Route desync (#4243)
Largely fixes a variety of trade route related desyncs, such as ones caused by different paths etc.

TR Cache was being updated only on the owner's client. Need to update on all or else route details could be different leading to desyncs. This is now done by invalidating the cache when the establish trade route unit mission is received.

I suspect there is still the potential for desyncs if the user opens the TR popup but something changes before they issue the command. Even just opening the popup could cause problems. I think a net message is in order for a real fix.