fix bug where plot reveal was dependant on active player
Return value from CvPlot::setRevealed is more nuanced than simply
returning whether the plot is being revealed to the team.

Now using a simple check.

This bug causes all sorts of problems, particularly desyncs in MP.
While the position is possibly synced with host, the visibilty of the
plots is not. This means things like barb camps get spawned in different
locations on different machines. This will be responsible for a large
number of desyncs for people who autoexplore.

Also, it causes the AI to explore suboptimally.

Will result in slightly more processing but it is the right thing to do.