Opinion Weight Fix
Fixed a small imbalance in the AI's calculation of opinion ranges, so that favorable and unfavorable opinions are now on an equal scale.

[Current Ranges]
150+: Unforgivable
70 to 149: Enemy
20 to 69: Competitor
-20 to 19: Neutral
-70 to -21: Favorable
-150 to -71: Friend
-151 or lower: Ally

[Fixed Ranges]
150+: Unforgivable
70 to 149: Enemy
20 to 69: Competitor
-19 to 19: Neutral
-69 to 20: Favorable
-149 to -70: Friend
-150 or lower: Ally

Also fixed a bug in the DoEstimateOtherPlayerOpinions function which set all positive opinions to FAVORABLE rather than using the weight properly and brought it in line with the fix above.