Allow AI to make & use improved Approach guesses
Rewrites the DoUpdateApproachTowardsUsGuesses function to allow the AI to make an educated guess as to another player's approach towards them.

Adds a new function to access the memory value for the AI's guess, GetTrueApproachTowardsUsGuess, since GetApproachTowardsUsGuess is used to show the AI's visible approach.

When a human player makes a demand or chooses the insulting dialogue option, the AI will now guess that their approach is HOSTILE.