Add Teammate Text Key
Makes the modifier "We are on the same team." be displayed at the top of the opinion modifiers list for teammates; does not actually change opinion score, but I think it's worth adding - it reassures players of the AI's allegiance towards them, and it's worth noting as a modifier that impacts Opinion - since AIs always consider teammates to be at the ALLY level.

The opinion penalty for asking a teammate not to settle near you appears in the mood display even though it doesn't actually impact Opinion, since it's set to ALLY. I set it to 1 if the players are teammates, so that players will still see it - indicating that the request was made, but it's dark red instead of bright red, and Transparent Diplomacy will show it as (-1) rather than (-20), as in actuality it is a non-factor for teammates.

Also fixes a typo in the below 50% approval rating text key.