Improve AI Opinion Estimates & Bugfix
Adds original capital capturing, ideologies, social policies, and the diplo bonus for a master towards a vassal to the AI's DoEstimateOtherPlayerOpinions function. Working on adding more to this soon.

Fixes a bug in GetPolicyScore that was preventing iNumPolicies from being factored in to the opinion weight, as appears to be intended (needed to flip the sign).