bHideTrueFeelings Approach Valuation Fix
All instances of bHideTrueFeelings being set to true when obtaining the AI's approach for deal valuation were removed recently.

As such, this adds a case for WAR and DECEPTIVE in each instance where they're missing.

In these added cases, DECEPTIVE is always set to the default value for the switch statement; WAR is set to INT_MAX where HOSTILE and GUARDED also are, and to the default value otherwise.