Add Coop War Warning Logic & Fix Bug
When I raised issue #5423, it was pointed out that the reason the AI had diplo inequalities when it came to the penalties for denying a coop war request was because the AI's choice to warn the other AI was largely RNG.

This commit fixes #5423 by equalizing the approach/opinion penalties between humans and the AI, and adding the function IsCoopWarRequestUnacceptable to determine whether or not to warn the target AI, with smarter logic and much less RNG; feel free to build on this decision-making process.

Also fixes what seems to be a bug in the DoSendStatementToPlayer function where the AI would never warn the target AI when asked by an AI to start a coop war, and cleaned up the confusing syntax which set the CoopWarState.