Bugfix
Fix a typo; having a DoF, DP or ALLY opinion with the asker should return false, not true, for IsCoopWarRequestUnacceptable.

Removed unnecessary variable in DoUpdateApproachTowardsUsGuesses.

Opinion modifier for asking a teammate not to settle nearby should now be 0, but still displayed in the mood table.

Slight code optimization to policy score calculation.