National Wonders Typo
Removed "City must have X building in the City", because already present.

Also, I noticed the "Circus Maximus" National Wonder has extra space (current version), which I found that is because of the space between the full point and "[NEWLINE]", so I fixed in multiple locations this possible creator of issues.

So now is like this ".[NEWLINE][NEWLINE]", without the space between "." and "[NEWLINE]".