Add function to calculate GPT loss from DoW
There doesn't seem to be a function which accurately calculates this (unless you count the woefully inadequate vanilla check in GetCoopWarScore), so I made one.

Will be useful in the event that a sanity check is added for the AI to avoid entering wars that'll harm their economy too much.

bool bOtherPlayerEstimate allows for ePlayer to call the function to see how much GPT the other player would lose in the event of war, but doesn't use information (TR/deal details) that ePlayer wouldn't have access to.