Improve Diplo Mechanics for Stolen Artifacts (adds new memory value)
AI will no longer give Open Borders to players who have stolen artifacts from them, if there are still artifacts in their lands that can be stolen by the other player.

AI will now always have an opinion penalty for stealing artifacts, regardless of if you make a promise or not:
1 artifact: +20 opinion
2 artifacts: +30 opinion
3+ artifacts: +40 opinion

Decay rate: AI forgets one artifact theft every 50 turns, and if 1 to 3 artifacts were stolen, the penalty will be reduced by 5 every 25 turns.

Making a promise and not breaking it reduces the penalty by 10.

Modified DoCounters so that the penalty decays, but does not completely disappear until the AI has confronted the player about it (except permanent war/multiplayer). This is to record the promise/refusal and to block war-related exploits that would bypass the Open Borders refusal.

Fixed diplo AI function call for City-States and Barbarians.