Small Combat Panel Fix Part 2
Same as Small Combat Panel Fix from above to make sure that doesn't overflow, which is caused by the two rows of the interception string.

Now for cities and present aa unit will show the intercept message, max two bonuses and then if there are more bonuses; the misc. one.