Update CvUnit.cpp
Changes only to  CvUnit::AI_promote  and  CvUnit::AI_promotionValue. 
Lots of rearrangement and new system for promotion value which should be easier to understand and modify later on.
I left comments where I wasn't sure about things and found a couple of small mistakes/weirdness from other files but the main thing I didn't know is when the #if defined were necessary. This meant I left the amphibious promotion as it was (it will be much lower valued than others as it is).