Rename CvSyncObject to CvSyncArchive because it now derives from Fraxis' interface
Add m_syncVars to CvSyncArchive. This is a pointer to distinct instances of Fraxis' FAutoVariable interfaces because the engine requires them for multiplayer. They are only allocated in multiplayer in order to save memory

Remove sync writer dispatch because the engine wants the data send through its own interfaces anyway

Deltas are now managed by Fraxis' interface because the delta table is a required part of the interface