AI fixes for Great Admiral Directives
Currently, the AI expends all of its Great Admirals (at the same turn) for luxuries when it has more than one, even when it's happy, as long as it's at peace. Considering that a human player would rarely expend GAs when happy  -except when they really have a lot-, and would instead prefer to keep a sizable reserve of GAs, I suggest these changes as a start.