A lot of religion and pantheon benefits ask for, in order: eReligion = GetFounderBenefitsReligion() if (eReligion == NO_RELIGION) eReligion = GetReligionInMostCities() ...
GetReligionInMostCities() hasn't accepted pantheon as a religion since 2015. GetFounderBenefitsReligion() uses GetReligionCreatedByPlayer, which was changed in the recent cleanup, so I'm rewriting it to include pantheons again.
we can figure out how to deal with this kind of code section later.

WIP - need the for-loop to use the pantheon if a religion isn't found. Pantheon always comes first, right

- Fixing Celts so that cities with their pantheon also get the 3 faith after they found a religion.

- RemoveOtherReligions loops through the religions less often