Human player handicap yields fix
The handicap yields for the human player were not being given due to the lack of the <HandicapInfo_DifficultyBonus> info. Added and copied the parameters used for the AI to the human player, plus two comments to reduce the odds of confusion between human and AI parameters.

Tested locally on emperor difficulty and working.