Rework of route evaluation
Route evaluation is now done in a reasonable manner.

Iroquois and Songhai will now not build both a railroad and a regular road to connect a pair of cities.

Individual plot route evaluations are now done more coherently with other plots that are on the same route(s).

Building unconnected railroads to get extra village bonuses is now handled in one place.