Base Commit
Added the new belief in the xml, its changes in pantheonchanges sql, and the text in newbelieftext xml

Also made the changes to dagda and bran that changed their names and move all healing/defense onto one of them.

Additional balance changes, some in the proposal, some modified. I.e. it was clear the new snow pantheon's numbers were not balanced (+2 yields on every tundra totally insane, for example).

https://forums.civfanatics.com/threads/6-66-revision-of-the-celtic-pantheons.686198/#post-16522823