Celts New Pantheon and minor balance changes (#10847)
* Base Commit

Added the new belief in the xml, its changes in pantheonchanges sql, and the text in newbelieftext xml

Also made the changes to dagda and bran that changed their names and move all healing/defense onto one of them.

Additional balance changes, some in the proposal, some modified. I.e. it was clear the new snow pantheon's numbers were not balanced (+2 yields on every tundra totally insane, for example).

https://forums.civfanatics.com/threads/6-66-revision-of-the-celtic-pantheons.686198/#post-16522823

* Azum check and balance pass

Fixed xml text hanging around at the end
swapped tags for dagda and bran so they make sense when reading the code

reverted +5 border to +3 border on epona but made it give production as well, all 4 yields reduced to +8 (so 32 overall versus 30). Should be fairly neutral to yields while reducing overlap!

Lugh, reverted 15% -> 10% on wonder building to not encroach on Egypt. Added yields on wonders to mimic Goddess of Beauty. Overall this seems /seemed like the worst one and hopefully this gives it some identity?

Changed +1 production on ogma to +1 culture to be more in the spirit of the pantheon and avoid overlap with Korea.

(new) Dagda gave +30 yields on birth, same as Mandir. Seems too weak so buffed it 20% (12 yields each).

* missed epona +3

fix

* Database check

accidentally a word on Lugh. Otherwise no errors.