New Difficulty Changes Structure
Removes DifficultyChanges.sql, separating it into one file for the Ancient Ruins changes and five files for the triggered instant yields, one file for each difficulty.

This should now be way easier to modify for the end user compared to having to do math. And it's easy enough to read from the file if someone wants to do difficulty bonus related proposals.

Adds backwards compatibility to some renamed but previously used columns in the HandicapInfos table - if these normally unused columns are used by a modmod, their effects will stack with those of the new columns.