Add functionality for homeland AI to distribute its workers
When an AI civilization becomes non-contiguous, the workers will previously be stuck in the region they are currently in, meaning that newly founded colonies will not get any tiles improved.

This commit adds functionality for the homeland AI to compute how many contiguous regions it has and how many improvements are needed to be built in them, and then distribute its workers among them.

Since this behavior may put the workers in danger, it will not be used for human automated workers, they will have to manually move their workers to their colonies.