Text updates
* Improvement text

Updated text descriptions for various improvements, including Aplekton, Oppidum, Obelisk, Villa, and Isibaya, to provide detailed information on their functions and historical context.

* dont mention Egypt UA in Obelisk text

Was trying to be helpful, but on second thought it might be confusing.

* Garden icon remove from classes

BuildingChanges are class changes, needs to be just base building

* separate Garden entry

like Museum

* Update Villa text

describes the tile claiming effect