Reduce Bandeirante Yields From Scouting From 3 To 2
Bandeirante yields from scouting are way too high and generally push Brazil to minimize its scouting until they are unlocked. On a standard map, they can give around 2-3 policies and 8-12 techs in quite a short period of time for exploring the rest of the map. This change makes it so they are strong but don't give a crazy amount of gold/science/culture.